pyproject.toml
[tool.poetry]
name = "s3-direct-upload"
version = "0.1.0"
description = ""
authors = ["Mikhail Porokhovnichenko <marazmiki@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7 || ^3.8 || ^3.9 || ^3.10"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pytest-isort = "^2.0.0"
pytest-flake8 = "^1.0.7"
requests = "^2.26.0"
python-dotenv = "^0.19"
boto3 = "^1.18.55"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
